use %lu instead of casting to uint

+++ gtkgame.c
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: gtkgame.c,v 1.663 2007/12/15 20:36:56 c_anthon Exp $
+ * $Id: gtkgame.c,v 1.664 2007/12/17 14:05:45 c_anthon Exp $
  */
 
 #include "config.h"
@@ -5566,7 +5566,7 @@ extern void SetRollouts( gpointer p, guint n, GtkWidget *pwIgnore )
 
     if( abs(rw.rcRollout.nSeed) != abs(rcRollout.nSeed) ) {
 	    /* seed may be unsigned long int */
-      sprintf( sz, "set rollout seed %u", (unsigned int)rw.rcRollout.nSeed );
+      sprintf( sz, "set rollout seed %lu", rw.rcRollout.nSeed );
       UserCommand( sz );
     }
